Removed temp mesh class, moved camera to player folder
