+++ src/core/player/Camera.cpp
@@ -1,4 +1,4 @@
-#include "core/camera/Camera.h"
+#include "core/player/Camera.h"
 
 Camera::Camera(glm::vec3 position, glm::vec3 up, float yaw, float pitch) 
     : position(position), front(glm::vec3(0.0f, 0.0f, -1.0f)),
